+++ src/main/java/com/example/kkm_back/Controller/PostController.java
@@ -2,14 +2,13 @@ package com.example.kkm_back.Controller;
 
 
 import com.example.kkm_back.Domain.Post;
-import com.example.kkm_back.Repository.PostDAO;
+import com.example.kkm_back.DAO.PostDAO;
 import com.example.kkm_back.Repository.PostRepository;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Map;
 
 @CrossOrigin
 @RestController
