+++ src/main/resources/mappers/post.xml
@@ -5,7 +5,7 @@
 <!-- 위 코드가 mapper 설정을 하는 코드 -->
 
 <!-- Mapper를 사용하는 Repository의 전체 경로를 적어줘야함. -->
-<mapper namespace="com.example.kkm_back.Repository.PostDAO">
+<mapper namespace="com.example.kkm_back.DAO.PostDAO">
     <insert id="postInsert" parameterType="com.example.kkm_back.Domain.Post" useGeneratedKeys="true" keyProperty="post_id" keyColumn="post_id">
         <selectKey resultType="int" keyProperty="post_id" order="AFTER">SELECT LAST_INSERT_ID()</selectKey>
         Insert into
